Apt install erlang on Linux CI.

+++ ci/brew-util.inc.sh
@@ -75,16 +75,23 @@ function brew_upgrade() {
 }
 
 # install erlang without wxmac bloat
-function brew_install_erlang() {
-    echo_do "brew: Installing erlang, without wxmac..."
-    brew tap linuxbrew/xorg
-    # using a for loop because 'xargs -r' is not part of the BSD version (MacOS)
-    # comm -23 <(brew deps erlang) <(brew deps wxmac) | sed "/^wxmac$/d" | xargs -r -L1 brew install
-    for FORMULA in $(comm -23 <(brew deps erlang) <(brew deps wxmac) | sed "/^wxmac$/d"); do
-        brew install ${FORMULA}
-    done
-    brew install --force erlang --ignore-dependencies || brew link --force --overwrite erlang
-    echo_done
+function install_erlang() {
+    if [[ "$OS" = "linux" ]] && [[ "${FORCE_BREW:-}" != "true" ]]
+    then
+        echo_do "apt: Installing erlang..."
+        apt_install erlang
+        echo_done
+    else
+        echo_do "brew: Installing erlang, without wxmac..."
+        brew tap linuxbrew/xorg
+        # using a for loop because 'xargs -r' is not part of the BSD version (MacOS)
+        # comm -23 <(brew deps erlang) <(brew deps wxmac) | sed "/^wxmac$/d" | xargs -r -L1 brew install
+        for FORMULA in $(comm -23 <(brew deps erlang) <(brew deps wxmac) | sed "/^wxmac$/d"); do
+            brew install ${FORMULA}
+        done
+        brew install --force erlang --ignore-dependencies || brew link --force --overwrite erlang
+        echo_done
+    fi
 }
 
 function brew_install() {
@@ -97,7 +104,7 @@ function brew_install() {
 
         case ${NAME} in
             erlang)
-                brew_install_erlang
+                install_erlang
                 continue
                 ;;
             *)
